solid list view state
